# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Klard – Booking platform for consultants
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /providers:
    get:
      operationId: listProviders
      tags: [providers]
      summary: Search/list providers
      parameters:
        - name: zip
          in: query
          schema: { type: string }
        - name: city
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
        - name: q
          in: query
          schema: { type: string }
        - name: minPrice
          in: query
          schema: { type: number }
        - name: maxPrice
          in: query
          schema: { type: number }
        - name: limit
          in: query
          schema: { type: integer }
        - name: offset
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: List of providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProviderSummary"
    post:
      operationId: createProvider
      tags: [providers]
      summary: Create a provider profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
  /providers/{id}:
    get:
      operationId: getProvider
      tags: [providers]
      summary: Get provider details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Provider detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
        "404":
          description: Not found
    patch:
      operationId: updateProvider
      tags: [providers]
      summary: Update provider profile
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
  /providers/me:
    get:
      operationId: getMyProviderProfile
      tags: [providers]
      summary: Get the authenticated user's provider profile
      responses:
        "200":
          description: Provider profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
        "404":
          description: Not found (not yet a provider)
  /providers/{id}/services:
    get:
      operationId: listProviderServices
      tags: [services]
      summary: List services for a provider
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Add a service to a provider
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    patch:
      operationId: updateService
      tags: [services]
      summary: Update a service
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
  /providers/{id}/reviews:
    get:
      operationId: listProviderReviews
      tags: [reviews]
      summary: List reviews for a provider
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
  /reviews:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Leave a review for a completed booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
...
components:
  schemas:
...
    ProviderSummary:
      type: object
      required: [id, displayName, category, city, zip, rating, reviewCount, minPrice]
      properties:
        id:          { type: integer }
        displayName: { type: string }
        category:    { type: string }
        categorySlug: { type: string }
        city:        { type: string }
        zip:         { type: string }
        rating:      { type: number }
        reviewCount: { type: integer }
        minPrice:    { type: number }
        bio:         { type: ["string", "null"] }
        avatarUrl:   { type: ["string", "null"] }
        logoUrl:     { type: ["string", "null"] }
        yearsExperience: { type: ["integer", "null"] }
        verified:    { type: boolean }
        subscriptionTier: { type: string, enum: [basic, premium] }
        requiresDirectBilling: { type: boolean }
    Provider:
      type: object
      required: [id, clerkUserId, displayName, category, city, zip, rating, reviewCount]
      properties:
        id:           { type: integer }
        clerkUserId:  { type: string }
        displayName:  { type: string }
        email:        { type: string }
        phone:        { type: ["string", "null"] }
        bio:          { type: ["string", "null"] }
        category:     { type: string }
        categorySlug: { type: string }
        city:         { type: string }
        zip:          { type: string }
        address:      { type: ["string", "null"] }
        website:      { type: ["string", "null"] }
        avatarUrl:    { type: ["string", "null"] }
        logoUrl:      { type: ["string", "null"] }
        yearsExperience: { type: ["integer", "null"] }
        rating:       { type: number }
        reviewCount:  { type: integer }
        verified:     { type: boolean }
        subscriptionTier: { type: string, enum: [basic, premium] }
        requiresDirectBilling: { type: boolean }
        premiumSince:    { type: ["string", "null"] }
        icalToken:       { type: ["string", "null"] }
        calendarSyncUrl: { type: ["string", "null"] }
        createdAt:    { type: string }
    ProviderInput:
      type: object
      required: [displayName, category, city, zip]
      properties:
        displayName:  { type: string }
        bio:          { type: string }
        category:     { type: string }
        city:         { type: string }
        zip:          { type: string }
        address:      { type: string }
        phone:        { type: string }
        website:      { type: string }
        logoUrl:      { type: string }
        yearsExperience: { type: integer }
    ProviderUpdate:
      type: object
      properties:
        displayName:  { type: string }
        bio:          { type: string }
        category:     { type: string }
        city:         { type: string }
        zip:          { type: string }
        address:      { type: string }
        phone:        { type: string }
        website:      { type: string }
        avatarUrl:    { type: string }
        logoUrl:      { type: string }
        yearsExperience: { type: integer }
    Service:
      type: object
      required: [id, providerId, name, price, durationMinutes, vatRate]
      properties:
        id:              { type: integer }
        providerId:      { type: integer }
        name:            { type: string }
        description:     { type: ["string", "null"] }
        price:           { type: number }
        netPrice:        { type: ["number", "null"] }
        vatRate:         { type: number }
        durationMinutes: { type: integer }
        createdAt:       { type: string }
    ServiceInput:
      type: object
      required: [name, durationMinutes]
      properties:
        name:            { type: string }
        description:     { type: string }
        price:           { type: number }
        netPrice:        { type: number }
        vatRate:         { type: number }
        durationMinutes: { type: integer }
    ServiceUpdate:
      type: object
      properties:
        name:            { type: string }
        description:     { type: string }
        price:           { type: number }
        netPrice:        { type: number }
        vatRate:         { type: number }
        durationMinutes: { type: integer }
...
    Review:
      type: object
      required: [id, bookingId, customerId, providerId, rating, createdAt]
      properties:
        id:           { type: integer }
        bookingId:    { type: integer }
        customerId:   { type: string }
        customerName: { type: ["string", "null"] }
        providerId:   { type: integer }
        rating:       { type: integer }
        comment:      { type: ["string", "null"] }
        createdAt:    { type: string }
    ReviewInput:
      type: object
      required: [bookingId, rating]
      properties:
        bookingId:  { type: integer }
        rating:     { type: integer }
        comment:    { type: string }
...