# lib/api-zod/src/generated/types/provider.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Klard – Booking platform for consultants
 * OpenAPI spec version: 0.1.0
 */
import type { ProviderSubscriptionTier } from "./providerSubscriptionTier";

export interface Provider {
  id: number;
  clerkUserId: string;
  displayName: string;
  email?: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  bio?: string | null;
  category: string;
  categorySlug?: string;
  city: string;
  zip: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  yearsExperience?: number | null;
  rating: number;
  reviewCount: number;
  verified?: boolean;
  subscriptionTier?: ProviderSubscriptionTier;
  requiresDirectBilling?: boolean;
  /** @nullable */
  premiumSince?: string | null;
  /** @nullable */
  icalToken?: string | null;
  /** @nullable */
  calendarSyncUrl?: string | null;
  createdAt?: string;
}