# lib/api-zod/src/generated/types/providerUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Klard – Booking platform for consultants
 * OpenAPI spec version: 0.1.0
 */

export interface ProviderUpdate {
  displayName?: string;
  bio?: string;
  category?: string;
  city?: string;
  zip?: string;
  address?: string;
  phone?: string;
  website?: string;
  avatarUrl?: string;
  logoUrl?: string;
  yearsExperience?: number;
}